fix(ga): stop placing the goal inside buildings
Buildings sit on a lattice 30 units apart starting at the grid origin. The
goal was positioned by arithmetic in its own local space that landed it 10
units from a building centre. Buildings reach up to 8 units from their
centre and the goal reaches 10 from its own, so up to 8 units of the goal
was buried inside a building.

The comment claimed this placement made the target "always reachable
despite roadblocks", which is backwards: being inside a building is what
makes it hard to reach.

Placement now derives from the grid, putting the goal halfway between two
building centres, in the middle of a road. That moves it from 10 to 15 units
from the nearest building centre, and from 2 to 7 units of clearance from
the widest building's surface.

Also makes the roadblock count settable from the command line, so the city
can be varied for a headless run without editing the shared ProceduralParam
asset, which would otherwise persist the change to disk.

For reference, 40 roadblocks blocks about 13% of the 312 road segments in a
13x13 city.

Widens totalDistanceCovered to a double. As a float it stopped being able to
add to its own running total at 2^24, and long runs pass that: the reported
distance simply froze at 16,777,216 units.

+++ Assets/Scripts/Genetic Algorithm/Manager.cs
@@ -55,7 +55,10 @@ public class Manager : MonoBehaviour
     public int resetStepCount = 25000;
 
     [Header("Save Values")]
-    public float totalDistanceCovered = 0f;
+    // double, not float: a float stops being able to add small values to a large
+    // running total at 2^24, and long runs pass that. The reported distance simply
+    // stopped increasing at 16,777,216 units.
+    public double totalDistanceCovered = 0d;
     public int numberOfCollisions = 0; // Number of collisions overall
     public int numberOfGoals = 0; // Number of goals reached
     public float maximumFitness = 0f;
